refactor: logging: split logging-related elements in different dirs
This commit creates a specific directory to logging structs/elements.
Inside the dir, there are 4 files: mod.rs contains just the module names;
log_on_error.rs contains the macro; garbage_collector.rs contains the old
log_gc mod and finally logger.rs, which contains the Logger impl and related
structs.
