+++ src/app/logging/mod.rs
@@ -1,0 +1,3 @@
+pub mod garbage_collector;
+pub mod log_on_error;
+pub mod logger;
